fs: error if a symlink points at a file that is not included in the snapshot
This implements @fd0's first idea here:
<https://github.com/restic/restic/issues/542#issuecomment-328263959>.

> First, I think it may be a good idea to print a warning message when a
> symlinks is saved and the target of the symlink exists and is not
> included in the backup. This way, users will know that some data
> referenced in the snapshot is not available upon restore.

Which I wholeheartedly agree with.

In the interest of keeping restic's cli simple, and keeping people's
data safe, I've opted to not make this configurable. I suppose you could
call this a breaking change, but I personally consider it a fix: restic
shouldn't claim it has successfully backed up a directory unless it can
actually recreate the structure of that directory. IMO, it's better to
fail-fast than to claim success, only to greatly disappoint someone
later on.